mouni

+++ src/app/inventory/inventory.component.ts
@@ -63,22 +63,37 @@ export class InventoryComponent implements OnInit {
  
  //This is for product delete
  delete(prodId:any){  
-   console.log(Number)
-     fetch("http://localhost:2000/products/deleteproduct/" + prodId,{
-      method:'DELETE',
-      headers:{
-        "access-Control-Allow-Origin":"*"
-      },
-     })
-     .then(response => response.json())
-     .then(result=>{
-      console.log(result)
-      this.getProduct()
-      alert('ok')
-    })     
-     .catch(err =>
-      console.log(err))    
-  } 
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed == true) {
+      fetch("http://localhost:2000/products/deleteproduct/" + prodId,{
+        method:'DELETE',
+        headers:{
+          "access-Control-Allow-Origin":"*"
+        },
+       })
+       .then(response => response.json())
+       .then(result=>{
+        console.log(result)
+      })     
+       .catch(err =>
+        console.log(err))    
+    } 
+      Swal.fire(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success'
+      )
+  })
+}
+    
 
   //this is for edit the product
   edit(products:any){
